branch-newui-refactor: convert to int for display, the precision of float is not needed here.
git-svn-id: https://svn.code.sf.net/p/jaer/code/branches/newui-refactor@4016 b7f4320f-462c-0410-a916-d9f35bb82d52

+++ src/net/sf/jaer2/viewer/jAER2joglpxfxcanvas.java
@@ -186,9 +186,9 @@ public class jAER2joglpxfxcanvas extends Application implements GLEventListener 
 					fpsFXTxt.setText("JOGL -> JavaFX FPS: " + fpsFXPrint);
 
 					fpsFXpftTxt.setText(String.format(
-						"JavaFX (perfTracker):\n\tavgFPS %f, instaFPS %f\n\tavgPulse %f, instaPulse %f",
-						perfTracker.getAverageFPS(), perfTracker.getInstantFPS(), perfTracker.getAveragePulses(),
-						perfTracker.getInstantPulses()));
+						"JavaFX (perfTracker):\n\tavgFPS %d, instaFPS %d\n\tavgPulse %d, instaPulse %d",
+						(int) perfTracker.getAverageFPS(), (int) perfTracker.getInstantFPS(),
+						(int) perfTracker.getAveragePulses(), (int) perfTracker.getInstantPulses()));
 
 					final Runtime rt = Runtime.getRuntime();
